% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{compute_msgc}
\alias{compute_msgc}
\title{Corrected genetic mating success}
\usage{
compute_msgc(fertilized_eggs)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function)}
}
\value{
List containing corrected MSg (MSgc) for females and males
}
\description{
Compute corrected genetic mating success MSgc using F. Rousset's metric
For the moment, this function is only applicable to all the seeds produced (i.e. sampling_groundtruth function, no sampling).
}
\details{
Compute genetic mating success for males and females with a correction to avoid a possible artificial correlation between RS/MSg.
}
