% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{compute_Q}
\alias{compute_Q}
\title{idem, je te laisse compléter si nécessaire (pense à générer la doc avant aussi (build -> check))}
\usage{
compute_Q(focal_list, method = "min")
}
\description{
idem, je te laisse compléter si nécessaire (pense à générer la doc avant aussi (build -> check))
}
