% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{compute_Qfoc}
\alias{compute_Qfoc}
\title{Compute Qfoc}
\usage{
compute_Qfoc(focal_list)
}
\value{
Qfoc
}
\description{
Compute Qfoc needed for MSgc computation
}
\details{
Product of couple production
}
