% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{correlated_paternity}
\alias{correlated_paternity}
\title{Correlated paternity}
\usage{
correlated_paternity(fertilized_eggs)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function or any sampling_XXX() method) (no default)}
}
\value{
Vector of size n_female with $r_p$ for each female.
}
\description{
Compute detailed paternity correlations (i.e. by female $r_p$)
}
\details{
Return per female correlated paternity. See companion function pop_average_rp() for further processing.
}
