% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Descriptive stats on MS/RS}
\usage{
descriptive_stats(ms_rs_from_sampling)
}
\arguments{
\item{ms_rs_from_sampling}{Dataframe with detailled MS/RS (from sampling() function)}
}
\value{
Vector of size n_female with $r_p$ for each female.
}
\description{
Compute descriptive statistics on MS/RS for all sampling method (including groundtruth) for each replicate
}
\details{
Return per sex, per sampling method and per replicate mean and std for MSo, MSg and RSg.
}
