% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{eggs_abortion}
\alias{eggs_abortion}
\title{Eggs abortion}
\usage{
eggs_abortion(fertilized_eggs, aborded_fraction = 0)
}
\arguments{
\item{fertilized_eggs}{List of females' eggs with fathers identities (e.g. output from pollen_competition()) (no default)}

\item{aborded_fraction}{Fraction(s) of aborted eggs by females (see details) (default = 0)}
}
\value{
List of size n_females, each element being a vector of males IDs corresponding to fertilized eggs - after abortions
}
\description{
Discount fertilized eggs from females - simulates abortion
}
\details{
Return a same output as pollen_competition() function but after abortion (i.e. given fraction of eggs is discounted by female).
aborded_fraction can be either a float (in [0, 1] range) or a vector of size n_females giving fraction for each female.
}
