% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{gametes_drawing}
\alias{gametes_drawing}
\title{Get gametes per individual}
\usage{
gametes_drawing(
  n_females = 100,
  n_males = 100,
  mean_gamete_female = 50,
  ratio_gamete = 10,
  female_distrib = "poisson",
  male_distrib = "normal",
  female_distrib_params = NULL,
  male_distrib_params = 0.1
)
}
\arguments{
\item{n_females}{Number of females (default value : 100)}

\item{n_males}{Number of males (default value : 100)}

\item{mean_gamete_female}{Mean number of gametes by female (default value : 50)}

\item{ratio_gamete}{Ratio between mean number of gametes by male over mean number of gametes by male (e.g., if ratio_gamete = 10, males have 10x more gametes than females) (default value : 10)}

\item{female_distrib}{Distribution of gametes by female (default value : poisson)}

\item{male_distrib}{Distribution of gametes by male (default value : normal)}

\item{female_distrib_params}{If distribution for female needs more than the mean, provide additional arguments}

\item{male_distrib_params}{If distribution for female needs more than the mean, provide additional arguments (here, Coefficient of variation)}
}
\value{
List with two vectors (gam_female, gam_male)
}
\description{
Draw the number of gametes per individual.
}
\details{
Implemented distributions for males/females are 'poisson', 'normal' or 'constant' - others might be possible in the future.
Normal distribution allows user to specify larger variance (by specifying the coefficient of variation) than expected with the more constrained Poisson distribution. The standard deviation
of the distribution is computed as 'cv x mean', cv being the X_distrib_params parameters, and the obtained values are then rounded.
Note that the 'constant' behaviour is redondant - can be done directly from 'normal' distribution with a null variation coefficient.

Note: if used distribution can lead to negatives values, these will be set to zero before being returned.

Future: other distributions might be implemented
}
