% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{get_male_comp_values}
\alias{get_male_comp_values}
\title{Draw male competitive value}
\usage{
get_male_comp_values(n_males = 100, mean_comp = 100, sd_comp = 0, plot = FALSE)
}
\arguments{
\item{n_males}{Number of males (default value : 100)}

\item{mean_comp}{Mean comp. value (default value : 100)}

\item{sd_comp}{Variability in comp. values (default value : 0, i.e. no competition !)}

\item{plot}{Should the comp. value histogram be plotted ? (default False)}
}
\value{
Vector of competitive values
}
\description{
Produce a vector of competitive values
}
\details{
Simple implementation, male competitive values are drawn from a normal distribution.
After drawing, values are transformed such that all values are strictly above zero (i.e. x' = x + min(x) + 0.01).
This is needed (at least for now) because these value are used later as probabilities...
Be careful when using low mean and large variance (i.e. a case where some negatives values will be drawn), the mean after rescaling will be different.

Note: When sd_comp is fixed to zero, pollen doesn't compete !
}
