% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{get_sexual_selection_components}
\alias{get_sexual_selection_components}
\title{Mating and reproductive success from sampled seeds}
\usage{
get_sexual_selection_components(
  fertilized_eggs,
  sampled_fertilized_eggs,
  n_males,
  paternity_share = TRUE
)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function)}

\item{sampled_fertilized_eggs}{List of size n_females containing father for each fertilized eggs after samping (from any sampling functions)}

\item{n_males}{Number of males (no default)}

\item{paternity_share}{Should males RS be rescaled by total number of females eggs (default TRUE)}
}
\value{
List containing MSg & RS for females and males
}
\description{
Compute RS/MS (genetic - i.e. in the sens of 'from fertilized eggs')
}
\details{
Compute mating success and reproductive success (i.e. number of offspring) for males and females.
Paternity share is considered by default (see paternity_share argument), and females RS is always compute from all eggs (not sampled ones).
}
