% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{ms_obs}
\alias{ms_obs}
\title{Mating Success (obs)}
\usage{
ms_obs(pollen_repartition)
}
\arguments{
\item{pollen_repartition}{Matrix females X males with number of male gametes by females (e.g., from pollen_export() function) (no default)}
}
\value{
List with two vectors (mso_female, mso_male)
}
\description{
Mating success from 'observation' (i.e. where pollens land)
}
\details{
Mating success from 'observation' (i.e. where pollens land)
}
