% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{ms_rs_scaling}
\alias{ms_rs_scaling}
\title{Standardized MS/RS}
\usage{
ms_rs_scaling(samples)
}
\arguments{
\item{samples}{Dataframe similar to the one produce by sampling()}
}
\value{
Standardize MS/RS
}
\description{
Standardized MS/RS by sex, sampling method, sampling parameters and replicate
}
\details{
Standardize MS/RS - using x / mean(x) (see Lande ??)
}
