% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{pollen_competition}
\alias{pollen_competition}
\title{Pollen competition}
\usage{
pollen_competition(pollen_repartition, males_comp_values, gametes_by_female)
}
\arguments{
\item{pollen_repartition}{See base::sample()}

\item{males_comp_values}{See base::sample()}
}
\value{
List of size n_females, each element being a vector of males IDs corresponding to fertilized eggs
}
\description{
Pollen competition to access 'eggs'
}
\details{

}
