% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{pollen_export}
\alias{pollen_export}
\title{Pollen (male) export distribution}
\usage{
pollen_export(
  n_females = NULL,
  baseline_alpha = NULL,
  pollen_repartition = c(0.01),
  plot = FALSE,
  gametes_by_male = NULL
)
}
\arguments{
\item{n_females}{Number of females (default value : NULL)}

\item{baseline_alpha}{Baseline alpha values for Dirichlet distr. - vector of size n_females; see details (default value : NULL)}

\item{pollen_repartition}{How to handle pollen heterogeneity distribution See details. (default value : 0.01)}

\item{plot}{Should pollen repartition be plotted ? (default False)}

\item{gametes_by_male}{Vector of number of gametes by males (e.g., from gametes_drawing() function)}
}
\value{
A matrix (n_females x n_males) with pollen load
}
\description{
Generate a matrix (n_females x n_males) with pollen load using Dirichlet & Multinomial distributions.
}
\details{
Return the number of pollen grains by females (rows) exported by each male (columns).
The general idea being to draw one vector per-male (\eqn{p_i = Dir(\alpha_i)}, for i in \eqn{1:n_males}) of relative pollen export to females (using Dirichlet distribution),
then, to draw a realization of this vector using Multinomial distribution (see also Polya distribution).

The baseline_alpha (i.e. \eqn{\alpha_i = baseline_alpha} for all i) represents the alpha vector of the Dirichlet distribution (i.e. and represent how pollen from one male
is, on average, distributed among females), and can be tuned to e.g. represent attractiveness of females or any component
making pollen load more or less likely on a given female. If not provided, the n_females variable will be used
to create an alpha vector containing only ones. Note that it means that at least baseline_alpha or n_females should be
provided, and n_females will not be used if baseline_alpha is provided.

pollen_repartition argument allows specifying various heterogeneity schems for pollen repartition beyond among-females 'preferences' (baseline_alpha).
Two levels of heterogenity can be considered.
(i) heterogeneity level among females, (ii) heterogeneity level among males for among-females heterogeneity. (Oo glurp).
When \eqn{\alpha} is the same for all males (no heterogeneity among males), the level of among-female heterogeneity is driven by a fixed float provided to pollen_reparition
(e.g., 0.01 or 10). Generally speaking, the largest is that value, the most homogeneous is the among-female pollen export. Actually, that value is used to scale the baseline_alpha
vector (i.e. newalpha = baseline_alpha * pollen_repartition), thus be careful as effect of this value on heterogeneity will depends on the number
of females as well as on the baseline_alpha values ! (see Dirichlet distribution for expressions on variance). A more convenient implementation might be attemp in the future
to make this dependency less embarassing. (could use that value as sum(alpha) but will still have that dependency on n_female ? or
trying to have a given quantity constant somewhere, but which one ? nor clear whether it is possible nor how to do it).


Alternatively, heterogeneity levels might be different among males. In these cases, pollen_repartition could be a vector of size n_males, and
baseline_alpha will be scaled for each male depending on its pollen_repartition value. A last option, is to specify a list(mean=XX,sd=XX) for pollen_repartition,
values provided here will be used to draw a value for each male from a 10^N(mean, sd) distribution.

## Might have some improvements to do here...

Because details might appear unclear, it's a good practice to look at plot displayed by the function. One of them represents the among-females
pollen repartition for a subsample of males (here, 15 - might be added as a argument of the function...), and the other, the average number of females
on which pollens from one male land.
}
