% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{pollen_limitation}
\alias{pollen_limitation}
\title{Pollen limitation}
\usage{
pollen_limitation(gametes_by_female, fertilized_eggs)
}
\arguments{
\item{gametes_by_female}{Vector of number of gametes by females (e.g., from gametes_drawing() function) (no default)}

\item{fertilized_eggs}{List of females' eggs with fathers identities (e.g. output from pollen_competition()) (no default)}
}
\value{
Number of eggs not fertilized, percentage of eggs not fertilized, number of females with eggs unfertilized, percentage of female with eggs unfertilized
}
\description{
Estimating metrics about the level of pollen competition
}
\details{
Return details about the level of pollen limitation: how many eggs have not been fertilized, among how many females.
}
