% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{pop_average_rp}
\alias{pop_average_rp}
\title{Pop. average correlated paternity}
\usage{
pop_average_rp(correlated_paternity, n_males)
}
\arguments{
\item{correlated_paternity}{Vector of size n_females containing by female correlated paternity (no default)}

\item{n_males}{Number of male in the population}
}
\value{
List of two elements, the uncorrected pop. average correlated paternity and the corrected version.
}
\description{
Return population average correlated paternity - using Dorken & Perry method and potential correction.
}
\details{
Compute pop. average correlated paternity (following Dorken & Perry, 2017) and corrected version.
}
