% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{rp_calculation}
\alias{rp_calculation}
\title{Correlated paternity (internal function)}
\usage{
rp_calculation(female_seed_set)
}
\arguments{
\item{female_seed_set}{female seed set}
}
\description{
Correlated paternity (internal function)
}
