% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{sample_handmade}
\alias{sample_handmade}
\title{Modified version of base::sample}
\usage{
sample_handmade(x, size, prob = NULL)
}
\arguments{
\item{x}{See base::sample()}

\item{size}{See base::sample()}

\item{prob}{See base::sample()}
}
\value{
A vector of length size with elements drawn from x (if x is a vector), otherwise see details
}
\description{
(internal) Draw a sample from vector, some behaviors of base::sample() are modified
}
\details{
Can handle cases for which base::sample() is not made for. E.g., if x vector is empty or size is 0 or NA, return NULL. If length(x)==1, return x (base::sample would have return from 1:x)
In addition, the function check whether size is greater than length(x), in that case, size is reduced to length(x).
}
