% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{sampling}
\alias{sampling}
\title{Sampling}
\usage{
sampling(
  fertilized_eggs,
  n_males,
  methods = NULL,
  scaled = TRUE,
  mso = NULL,
  gametes = NULL,
  n_rep = 1
)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function) (no default)}

\item{n_males}{Number of males (no default)}

\item{methods}{List of lists containing methods to use and associated parameters (see examples)}

\item{scaled}{Should the MS/RS should be scaled ? (default TRUE)}

\item{mso}{MS obtained directly from observations (i.e. pollen repartition over females before pollen competition) (from ms_obs() function, not required)}

\item{gametes}{Number of gametes for each females/males (from gametes_drawing() function, not required)}

\item{n_rep}{Number of time each sampling method shoud be replicated (default 1)}
}
\value{
Dataframe with MS (obs), MS (gen), RS (gen), gamete counts, sex, sampling_method, parameters used for the sampling method, and replicate IDs
}
\description{
Convenient wrapper around sampling strategies with convenient output format
}
\details{
Run all required sampling methods with associated parameters n_rep time each. Results can be scaled by sex, methods & methods parameters.
User defined function can also be used here ; two constraints : (i) first two arguments must be 'fertilized_eggs' and 'n_males', others can be
anything defined by the user, (ii) the function must return list with four elements : msg_female, rsg_female, msg_male, rsg_male.
}
\examples{
\dontrun{
methods = list(base = list(method = "sampling_groundtruth",
                           params = list()),
               fixed = list(method = "sampling_fixed",
                            params = list(by_female_samples = 50, undercount_female = 'keep')),
               prorata = list(method = "sampling_prorata",
                              params = list(by_female_prop = 0.10, min_threshold = 0,
                                            undercount_female = 'keep', upsample_strategy = 's2')),
               random = list(method = "sampling_random",
                             params = list()))

sampling(fertilized_eggs, n_males, methods = methods, mso = mso, gametes = gametes, scaled = T)
}
}
