% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{sampling_fixed}
\alias{sampling_fixed}
\title{Fixed number sampling}
\usage{
sampling_fixed(
  fertilized_eggs,
  n_males,
  total_samples = 1000,
  by_female_samples = NULL,
  undercount_female = "remove"
)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function) (no default)}

\item{n_males}{Number of males (no default)}

\item{total_samples}{Total number of eggs to be sample (default 1000)}

\item{by_female_samples}{Number of eggs to be sample by female (no default)}

\item{undercount_female}{Which strategy to use when some females doesn't have enough eggs to be sample ? (see details, default 'remove')}
}
\value{
Sampled fertilized eggs
}
\description{
Samples females eggs using a fixed number (So call 'Fixed Strategy', see details)
}
\details{
Sample eggs using a 'fixed number' strategy (i.e. a fixed number of eggs by female is sampled). This number is determined using either
total_samples or by_female_samples argument (thus at least one should be provided if not using the default value. If both are provided,
the by_female_samples will be used).

In some cases, some females might not have enough eggs to sample. Two strategies can be used to consider these cases (i) 'remove' strategy,
(ii) 'keep' strategy. The default 'remove' strategy doesn't consider females w/o enough eggs (will be NA in the output), the 'keep' strategy
consider these females by sampling all their eggs.

In all cases, if target number of samples is not reached because some females haven't enough eggs, additional eggs are selected from others females (in a
uniform and random way).
}
