% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{sampling_groundtruth}
\alias{sampling_groundtruth}
\title{Sampling groundtruth}
\usage{
sampling_groundtruth(fertilized_eggs, n_males)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function) (no default)}

\item{n_males}{Number of males (no default)}
}
\value{
List containing MS & RS for females and for males
}
\description{
Not sampling per se, compute exact RS/MS (genetic - i.e. in the sens of 'from fertilized eggs')
}
\details{
(Should be kept as an internal function ?)

Compute exact mating success (i.e. number of 'genetic' mates) and reproductive success (i.e. number of offspring) for males and females
}
