% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{sampling_random}
\alias{sampling_random}
\title{Random sampling}
\usage{
sampling_random(fertilized_eggs, n_males, total_samples = 1000)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function) (no default)}

\item{n_males}{Number of males (no default)}

\item{total_samples}{Total number of eggs to be sample (default 1000)}
}
\value{
List containing MS & RS for females and for males
}
\description{
Samples eggs randomly (i.e. independently of female identity, see details)
}
\details{
(Should be kept as an internal function ?)

Compute mating success (i.e. number of 'genetic' mates) and reproductive success (i.e. number of offspring) for males and females
using a fully random strategy. A given number of eggs (total_samples) is randomly chosen, then parents identity is checked to compute MS/RS for males
et MS for females.

Note that in all cases, the RS for females is computed using all eggs ! (should add another option here to allow this behavior to be changed ?)
}
