% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_components.R
\name{get_male_comp_values}
\alias{get_male_comp_values}
\title{Draw male competitive value}
\usage{
get_male_comp_values(
  n_males = 100,
  distrib = sample,
  dist_params = list(x = 100, replace = TRUE),
  translation = TRUE,
  plot = FALSE
)
}
\arguments{
\item{n_males}{Number of males (default value : 100)}

\item{distrib}{Prob. distribution to use (default value : sample, i.e. uniform distrib.)}

\item{dist_params}{Distribution parameters}

\item{translation}{Should the resulting comp. value be translated to be > 0 ? (default TRUE)}

\item{plot}{Should the comp. value histogram be plotted ? (default FALSE)}
}
\value{
Vector of competitive values
}
\description{
Produce a vector of competitive values
}
\details{
Draw males competitive values from provided distribution (uniform by default).
After drawing, values are transformed such that all values are strictly above zero (i.e. x' = x + min(x) + 0.01).
This is needed because these value are used later as probabilities. (that default behavior can be modified with the translation arg.)
Note however this may change mean and variance of specified distribution !
}
