% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_strategies.R
\name{sampling_groundtruth}
\alias{sampling_groundtruth}
\title{Sampling ground-truth}
\usage{
sampling_groundtruth(fertilized_eggs, ...)
}
\arguments{
\item{fertilized_eggs}{List of size n_females containing father for each fertilized eggs (from e.g., pollen_competition() function) (no default)}
}
\value{
Sampled fertilized eggs
}
\description{
Dummy/Convenience function - just return the input
}
\details{
Return the input list - use for generality purpose, same pipeline can be use than with other sampling methods.
}
